[Desktop] Add tag assignment to file inspector
