+++ packages/ui/src/DropdownMenu.tsx
@@ -1,6 +1,13 @@
 import * as RadixDM from '@radix-ui/react-dropdown-menu';
 import clsx from 'clsx';
-import React, { PropsWithChildren, Suspense, useCallback, useRef, useState } from 'react';
+import React, {
+	PropsWithChildren,
+	Suspense,
+	useCallback,
+	useContext,
+	useRef,
+	useState
+} from 'react';
 import { Link } from 'react-router-dom';
 import {
 	ContextMenuDivItem,
@@ -15,6 +22,9 @@ interface DropdownMenuProps extends RadixDM.MenuContentProps {
 	triggerClassName?: string;
 	alignToTrigger?: boolean;
 }
+
+const context = React.createContext<boolean>(false);
+export const useDropdownMenu = () => useContext(context);
 
 const Root = ({
 	trigger,
@@ -43,7 +53,7 @@ const Root = ({
 					style={{ width }}
 					{...props}
 				>
-					{children}
+					<context.Provider value={true}>{children}</context.Provider>
 				</RadixDM.Content>
 			</RadixDM.Portal>
 		</RadixDM.Root>
